@@ -1,11 +1,16 @@
 <?php
 
 return [
-    'max_precipitation' => 0,
-    'max_air_temperature' => 26,
-    'min_air_temperature' => 14,
-    'min_air_temperature_if_clear' => 11,
-    'hours_to_check' => 4,
-    'place_code_to_check' => 'vilnius',
-    'max_wind_speed' => 8,
+    'rules' => [
+        'max_precipitation' => 0,
+        'max_air_temperature' => 26,
+        'min_air_temperature' => 14,
+        'min_air_temperature_if_clear' => 11,
+        'hours_to_check' => 4,
+        'max_wind_speed' => 8,
+        'max_wind_gust' => 12,
+    ],
+    'available_places' => [
+        'vilnius' => 'Vilnius',
+    ],
 ];
