@@ -6,5 +6,7 @@ return [
     'error' => 'The weather is not suitable for basketball: ',
     'too_high_air_temperature' => ':hour h to high air temperature: :airTemperature C.',
     'too_low_air_temperature' => ':hour h to low air temperature: :airTemperature C.',
-    'too_high_wind_speed' => ':hour h to high wind speed: :windSpeed m/s.'
+    'too_high_wind_speed' => ':hour h to high wind speed: :windSpeed m/s.',
+    'too_high_wind_gust' => ':hour h to high wind gust: :windGust m/s.',
+    'heavy_snow' => ':hour h heavy snow.',
 ];
