@@ -2,5 +2,6 @@
 
 return [
     'check_weather' => 'Check weather',
-    'check_weather_for_basketball' => 'Check weather for basketball'
+    'check_weather_for_basketball' => 'Check weather for basketball',
+    'select_place' => 'Please select place',
 ];
