@@ -2,10 +2,11 @@
 
 return [
     'phone' => [
-        'new_user_verification_code' => 'Sveiki, Jūsų registracijos kodas yra: {code}.',
+        'new_user_verification_code' => 'Labas! Tavo registracijos kodas yra: {code}.',
         'new_user_code_not_sent' => 'Registracijos kodas neišsiųstas, prašome pamėgint vėliau.',
         'code_verification_notice' => 'Prieš prisijungiant, prašome patikrinti telefoną ir įvesti gautą registracijos kodą.',
         'verify_phone' => 'Pativirtinti telefono numerį',
+        'success' => 'Telefono numeris sėkmingai patvirtintas.',
     ],
     'verify' => 'Patvirtinti',
     'verification_code' => 'Registracijos kodas',
