@@ -6,5 +6,7 @@ return [
     'error' => 'Oras netinkamas krepšiniui',
     'too_high_air_temperature' => ':hour valandą per aukšta oro temperatūra: :airTemperature C.',
     'too_low_air_temperature' => ':hour valandą per žema oro temperatūra: :airTemperature C.',
-    'too_high_wind_speed' => ':hour valandą per didelis vėjo greitis: :windSpeed m/s.'
+    'too_high_wind_speed' => ':hour valandą per didelis vėjo greitis: :windSpeed m/s.',
+    'too_high_wind_gust' => ':hour valandą per didelis vėjo gūsių greitis: :windGust m/s.',
+    'heavy_snow' => ':hour valandą stiprus sniegas.',
 ];
