@@ -14,8 +14,10 @@ use Illuminate\Support\Facades\Route;
 | is assigned the "api" middleware group. Enjoy building your API!
 |
 */
-
-Route::get('weather-warnings', [WeatherApiController::class, 'getWeatherWarnings']);
+Route::prefix('weather')->group(function () {
+    Route::get('warnings', [WeatherApiController::class, 'getWeatherWarnings']);
+    Route::get('available-places', [WeatherApiController::class, 'getAvailablePlaces']);
+});
 
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
