@@ -48,7 +48,8 @@ abstract class AbstractRequest implements StatsAwareRequestInterface
     {
         return static function (TransferStats $stats)
         {
-            Log::info(sprintf('Transfer time: %s, Uri: %s.', $stats->getTransferTime(), $stats->getEffectiveUri()));
+            $message = sprintf('Transfer time: %s, Uri: %s.', $stats->getTransferTime(), $stats->getEffectiveUri());
+            Log::channel('client')->info($message);
         };
     }
 }
